feat(list): FileType enum is now public; deprecated File::from...line in favour of `LineParser`
It is now possible to retrieve the `FileType` from file using `File::file_type()`.

closes #128
