+++ suppaftp/src/list.rs
@@ -7,43 +7,42 @@
 //! If you find a variant which doesn't work with this parser,
 //! please feel free to report an issue to <https://github.com/veeso/suppaftp>.
 //!
+//! This module also exposes the [`File`] type which represents a file entry on the remote system.
+//! You can distinguish whether the entry is a file, a directory or a symlink by checking the appropriate methods on the [`File`] type or
+//! by getting the [`FileType`] enum.
+//!
 //! ## Get started
 //!
 //! Whenever you receive the output for your LIST command, all you have to do is to iterate over lines and
-//! call `File::from_line()` function as shown in the example.
+//! call [`File::from_str`] function as shown in the example.
 //!
-//! ```rust,ignore
-//! use std::convert::TryFrom;
-//! use suppaftp::{FtpStream, list::File};
+//! ```rust
+//! use suppaftp::list::{File, ListParser};
 //!
-//! // Connect to the server
-//! let mut ftp_stream = FtpStream::connect("127.0.0.1:10021").unwrap_or_else(|err|
-//!     panic!("{}", err)
-//! );
+//! // imagine this line received from a LIST command
+//! let line = "-rw-r--r-- 1 user group 1234 Nov 5 13:46 example.txt";
 //!
-//! // Authenticate
-//! assert!(ftp_stream.login("test", "test").is_ok());
+//! let file = ListParser::parse_posix(line).expect("failed to parse LIST line");
 //!
-//! // List current directory
-//! let files: Vec<File> = ftp_stream.list(None).ok().unwrap().iter().map(|x| File::try_from(x.as_str()).ok().unwrap()).collect();
-//!
-//! // Disconnect from server
-//! assert!(ftp_stream.quit().is_ok());
-//!
+//! assert_eq!(file.name(), "example.txt");
 //! ```
 
-use std::convert::TryFrom;
+mod file;
+mod file_type;
+mod pex;
+
 use std::ops::Range;
-use std::path::{Path, PathBuf};
-use std::str::FromStr;
+use std::path::PathBuf;
 use std::time::{Duration, SystemTime};
 
-use chrono::Datelike;
-use chrono::prelude::{NaiveDate, NaiveDateTime, Utc};
+use chrono::prelude::Utc;
+use chrono::{Datelike, NaiveDate, NaiveDateTime};
 use lazy_regex::{Lazy, Regex};
 use thiserror::Error;
 
-// -- Regex
+pub use self::file::File;
+pub use self::file_type::FileType;
+pub use self::pex::{PosixPex, PosixPexQuery};
 
 /// POSIX system regex to parse list output
 static POSIX_LS_RE: Lazy<Regex> = lazy_regex!(
@@ -53,56 +52,10 @@ static POSIX_LS_RE: Lazy<Regex> = lazy_regex!(
 static DOS_LS_RE: Lazy<Regex> =
     lazy_regex!(r#"^(\d{2}\-\d{2}\-\d{2}\s+\d{2}:\d{2}\s*[AP]M)\s+(<DIR>)?([\d,]*)\s+(.+)$"#);
 
-// -- File entry
+/// Result type for parsing LIST lines
+pub type ParseResult<T> = Result<T, ParseError>;
 
-/// Describes a file entry on the remote system.
-/// This data type is returned in a collection after parsing a LIST output
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-pub struct File {
-    /// File name
-    name: String,
-    /// File type describes whether it is a directory, a file or a symlink
-    file_type: FileType,
-    /// File size in bytes
-    size: usize,
-    /// Last time the file was modified
-    modified: SystemTime,
-    /// User id (POSIX only)
-    uid: Option<u32>,
-    /// Group id (POSIX only)
-    gid: Option<u32>,
-    /// POSIX permissions
-    posix_pex: (PosixPex, PosixPex, PosixPex),
-}
-
-/// Describes the kind of file. Can be `Directory`, `File` or `Symlink`. If `Symlink` the path to the pointed file must be provided
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-enum FileType {
-    Directory,
-    File,
-    Symlink(PathBuf),
-}
-
-/// ### PosixPexQuery
-///
-/// This enum is used to query about posix permissions on a file
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub enum PosixPexQuery {
-    Owner,
-    Group,
-    Others,
-}
-
-/// Describes the permissions on POSIX system.
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash)]
-struct PosixPex {
-    read: bool,
-    write: bool,
-    execute: bool,
-}
-
-// -- Error
-
+/// Errors that can occur when parsing a LIST line
 #[derive(Debug, Error, Eq, PartialEq)]
 pub enum ParseError {
     #[error("Syntax error: invalid line")]
@@ -113,82 +66,52 @@ pub enum ParseError {
     BadSize,
 }
 
-impl File {
-    // -- getters
+/// Parser for a line of the `LIST` command output
+///
+/// You can use this parser to parse a [`File`] from a **line** returned by the `LIST` command.
+///
+/// You can use either the [`ListParser::parse`] which tries to automatically detect the format (POSIX or DOS)
+/// or you can use the more specific methods [`ListParser::parse_posix`] and [`ListParser::parse_dos`].
+///
+/// # Notes
+///
+/// The [`ListParser`] has since `7.1.0` replaced the [`File::from_posix_line`] and [`File::from_dos_line`] methods.
+pub struct ListParser;
 
-    /// Get file name
-    pub fn name(&self) -> &str {
-        self.name.as_str()
+impl ListParser {
+    /// Parse an output line from a MLSD command
+    /// Returns a [`File`] instance if parsing is successful, otherwise returns a [`ParseError`].
+    ///
+    /// MLSD syntax has the following syntax:
+    ///
+    /// ```text
+    /// type=dir;modify=20201019151930;UNIX.mode=0755;UNIX.uid=1000;UNIX.gid=1000; pub
+    /// ```
+    pub fn parse_mlsd(line: &str) -> ParseResult<File> {
+        Self::parse_mlsx(line)
     }
 
-    /// Get whether file is a directory
-    pub fn is_directory(&self) -> bool {
-        self.file_type.is_directory()
+    /// Parse an output line from a MLST command
+    /// Returns a [`File`] instance if parsing is successful, otherwise returns a [`ParseError`].
+    ///
+    /// MLST syntax has the following syntax:
+    ///
+    /// ```text
+    /// type=dir;modify=20201019151930;UNIX.mode=0755;UNIX.uid=1000;UNIX.gid=1000; pub
+    /// ```
+    pub fn parse_mlst(line: &str) -> ParseResult<File> {
+        Self::parse_mlsx(line)
     }
-
-    /// Get whether file is a file
-    pub fn is_file(&self) -> bool {
-        self.file_type.is_file()
-    }
-
-    /// Get whether file is a symlink
-    pub fn is_symlink(&self) -> bool {
-        self.file_type.is_symlink()
-    }
-
-    /// Returns, if available, the file the symlink is pointing to
-    pub fn symlink(&self) -> Option<&Path> {
-        self.file_type.symlink()
-    }
-
-    /// Returned file size in bytes
-    pub fn size(&self) -> usize {
-        self.size
-    }
-
-    /// Returns the last time the file was modified
-    pub fn modified(&self) -> SystemTime {
-        self.modified
-    }
-
-    /// Returns when available the owner user of the file. (POSIX only)
-    pub fn uid(&self) -> Option<u32> {
-        self.uid.to_owned()
-    }
-
-    /// Returns when available the owner group of the file. (POSIX only)
-    pub fn gid(&self) -> Option<u32> {
-        self.gid.to_owned()
-    }
-
-    /// Returns whether `who` can read file
-    pub fn can_read(&self, who: PosixPexQuery) -> bool {
-        self.query_pex(who).can_read()
-    }
-
-    /// Returns whether `who` can write file
-    pub fn can_write(&self, who: PosixPexQuery) -> bool {
-        self.query_pex(who).can_write()
-    }
-
-    /// Returns whether `who` can execute file
-    pub fn can_execute(&self, who: PosixPexQuery) -> bool {
-        self.query_pex(who).can_execute()
-    }
-
-    /// Returns the pex structure for selected query
-    fn query_pex(&self, who: PosixPexQuery) -> &PosixPex {
-        match who {
-            PosixPexQuery::Group => &self.posix_pex.1,
-            PosixPexQuery::Others => &self.posix_pex.2,
-            PosixPexQuery::Owner => &self.posix_pex.0,
-        }
-    }
-
-    // -- parsers
 
     /// Parse an output line from a MLSD or MLST command
-    pub fn from_mlsx_line(line: &str) -> Result<Self, ParseError> {
+    /// Returns a [`File`] instance if parsing is successful, otherwise returns a [`ParseError`].
+    ///
+    /// MLSD/MLST syntax has the following syntax:
+    ///
+    /// ```text
+    /// type=dir;modify=20201019151930;UNIX.mode=0755;UNIX.uid=1000;UNIX.gid=1000; pub
+    /// ```
+    fn parse_mlsx(line: &str) -> ParseResult<File> {
         let tokens = line.split(';').collect::<Vec<&str>>();
         if tokens.is_empty() {
             return Err(ParseError::SyntaxError);
@@ -264,9 +187,15 @@ impl File {
         Ok(f)
     }
 
-    /// Parse a POSIX LIST output line and if it is valid, return a `File` instance.
-    /// In case of error a `ParseError` is returned
-    pub fn from_posix_line(line: &str) -> Result<Self, ParseError> {
+    /// Parse a POSIX LIST output line and if it is valid, return a [`File`] instance, otherwise return a [`ParseError`].
+    ///
+    /// POSIX syntax has the following syntax:
+    ///
+    /// ```text
+    /// {FILE_TYPE}{PERMISSIONS} {LINK_COUNT} {USER} {GROUP} {FILE_SIZE} {MODIFIED_TIME} {FILENAME}
+    /// -rw-r--r-- 1 user group 1234 Nov 5 13:46 example.txt
+    /// ```
+    pub fn parse_posix(line: &str) -> ParseResult<File> {
         // Apply regex to result
         match POSIX_LS_RE.captures(line) {
             // String matches regex
@@ -358,7 +287,7 @@ impl File {
     }
 
     /// Try to parse a "LIST" output command line in DOS format.
-    /// Returns error if syntax is not DOS compliant.
+    /// Returns [`ParseError`] if syntax is not DOS compliant.
     /// DOS syntax has the following syntax:
     ///
     /// ```text
@@ -366,7 +295,7 @@ impl File {
     /// 10-19-20  03:19PM <DIR> pub
     /// 04-08-14  03:09PM 403   readme.txt
     /// ```
-    pub fn from_dos_line(line: &str) -> Result<Self, ParseError> {
+    pub fn parse_dos(line: &str) -> ParseResult<File> {
         // Apply regex to result
         match DOS_LS_RE.captures(line) {
             // String matches regex
@@ -483,168 +412,30 @@ impl File {
     }
 }
 
-impl FromStr for File {
-    type Err = ParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Self::try_from(s)
-    }
-}
-
-impl TryFrom<&str> for File {
-    type Error = ParseError;
-
-    fn try_from(line: &str) -> Result<Self, Self::Error> {
-        // First try to parse the line in POSIX format (vast majority case).
-        match Self::from_posix_line(line) {
-            Ok(entry) => Ok(entry),
-            // If POSIX parsing fails, try with DOS parser.
-            Err(_) => match Self::from_dos_line(line) {
-                Ok(entry) => Ok(entry),
-                Err(err) => Err(err),
-            },
-        }
-    }
-}
-
-impl TryFrom<&String> for File {
-    type Error = ParseError;
-
-    fn try_from(line: &String) -> Result<Self, Self::Error> {
-        Self::try_from(line.as_str())
-    }
-}
-
-impl TryFrom<String> for File {
-    type Error = ParseError;
-
-    fn try_from(line: String) -> Result<Self, Self::Error> {
-        File::try_from(line.as_str())
-    }
-}
-
-impl FileType {
-    /// Returns whether the file is a directory
-    fn is_directory(&self) -> bool {
-        matches!(self, &FileType::Directory)
-    }
-
-    /// Returns whether the file is a file
-    fn is_file(&self) -> bool {
-        matches!(self, &FileType::File)
-    }
-
-    /// Returns whether the file is a symlink
-    fn is_symlink(&self) -> bool {
-        matches!(self, &FileType::Symlink(_))
-    }
-
-    /// get symlink if any
-    fn symlink(&self) -> Option<&Path> {
-        match self {
-            FileType::Symlink(p) => Some(p.as_path()),
-            _ => None,
-        }
-    }
-}
-
-impl PosixPex {
-    /// Returns whether read permission is true
-    fn can_read(&self) -> bool {
-        self.read
-    }
-
-    /// Returns whether write permission is true
-    fn can_write(&self) -> bool {
-        self.write
-    }
-
-    /// Returns whether execute permission is true
-    fn can_execute(&self) -> bool {
-        self.execute
-    }
-}
-
-impl Default for PosixPex {
-    fn default() -> Self {
-        Self {
-            read: true,
-            write: true,
-            execute: true,
-        }
-    }
-}
-
-impl From<u8> for PosixPex {
-    fn from(bits: u8) -> Self {
-        Self {
-            read: ((bits >> 2) & 0x01) != 0,
-            write: ((bits >> 1) & 0x01) != 0,
-            execute: (bits & 0x01) != 0,
-        }
-    }
-}
-
 #[cfg(test)]
-mod test {
-
+mod tests {
     use chrono::DateTime;
-    use pretty_assertions::assert_eq;
 
     use super::*;
 
     #[test]
-    fn file_getters() {
-        let file: File = File {
-            name: String::from("provola.txt"),
-            file_type: FileType::File,
-            size: 2048,
-            modified: SystemTime::UNIX_EPOCH,
-            gid: Some(0),
-            uid: Some(0),
-            posix_pex: (PosixPex::from(7), PosixPex::from(5), PosixPex::from(4)),
-        };
-        assert_eq!(file.name(), "provola.txt");
-        assert_eq!(file.is_directory(), false);
-        assert_eq!(file.is_file(), true);
-        assert_eq!(file.is_symlink(), false);
-        assert_eq!(file.symlink(), None);
-        assert_eq!(file.size(), 2048);
-        assert_eq!(file.gid(), Some(0));
-        assert_eq!(file.uid(), Some(0));
-        assert_eq!(file.modified(), SystemTime::UNIX_EPOCH);
-        // -- posix pex
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), true);
-        assert_eq!(file.can_write(PosixPexQuery::Group), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), true);
-        assert_eq!(file.can_read(PosixPexQuery::Others), true);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), false);
-    }
-
-    #[test]
     fn parse_posix_line() {
-        let file: File = File::from_str("-rw-rw-r-- 1 0  1  8192 Nov 5 2018 omar.txt")
-            .ok()
-            .unwrap();
-        assert_eq!(file.name(), "omar.txt");
-        assert_eq!(file.size, 8192);
-        assert_eq!(file.is_file(), true);
-        assert_eq!(file.uid, Some(0));
-        assert_eq!(file.gid, Some(1));
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), false);
-        assert_eq!(file.can_read(PosixPexQuery::Group), true);
-        assert_eq!(file.can_write(PosixPexQuery::Group), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), false);
-        assert_eq!(file.can_read(PosixPexQuery::Others), true);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), false);
-        assert_eq!(
+        let file = ListParser::parse_posix("-rw-rw-r-- 1 0  1  8192 Nov 5 2018 omar.txt").unwrap();
+        pretty_assertions::assert_eq!(file.name(), "omar.txt");
+        pretty_assertions::assert_eq!(file.size, 8192);
+        pretty_assertions::assert_eq!(file.is_file(), true);
+        pretty_assertions::assert_eq!(file.uid, Some(0));
+        pretty_assertions::assert_eq!(file.gid, Some(1));
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), false);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(
             file.modified()
                 .duration_since(SystemTime::UNIX_EPOCH)
                 .ok()
@@ -652,24 +443,23 @@ mod test {
             Duration::from_secs(1541376000)
         );
         // Group and user as strings; directory
-        let file: File = File::from_str("drwxrwxr-x 1 root  dialout  4096 Nov 5 2018 provola")
-            .ok()
-            .unwrap();
-        assert_eq!(file.name(), "provola");
-        assert_eq!(file.size, 4096);
-        assert_eq!(file.is_directory(), true);
-        assert_eq!(file.uid, None);
-        assert_eq!(file.gid, None);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), true);
-        assert_eq!(file.can_write(PosixPexQuery::Group), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), true);
-        assert_eq!(file.can_read(PosixPexQuery::Others), true);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), true);
-        assert_eq!(
+        let file =
+            ListParser::parse_posix("drwxrwxr-x 1 root  dialout  4096 Nov 5 2018 provola").unwrap();
+        pretty_assertions::assert_eq!(file.name(), "provola");
+        pretty_assertions::assert_eq!(file.size, 4096);
+        pretty_assertions::assert_eq!(file.is_directory(), true);
+        pretty_assertions::assert_eq!(file.uid, None);
+        pretty_assertions::assert_eq!(file.gid, None);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(
             file.modified()
                 .duration_since(SystemTime::UNIX_EPOCH)
                 .ok()
@@ -677,104 +467,101 @@ mod test {
             Duration::from_secs(1541376000)
         );
         // Setuid bit
-        let file: File =
-            File::from_str("drws------    2 u-redacted g-redacted      3864 Feb 17  2023 sas")
-                .ok()
-                .unwrap();
-        assert_eq!(file.is_directory(), true);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), false);
-        assert_eq!(file.can_write(PosixPexQuery::Group), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), false);
-        assert_eq!(file.can_read(PosixPexQuery::Others), false);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), false);
-        let file: File =
-            File::from_str("drwS------    2 u-redacted g-redacted      3864 Feb 17  2023 sas")
-                .ok()
-                .unwrap();
-        assert_eq!(file.is_directory(), true);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), false);
-        assert_eq!(file.can_read(PosixPexQuery::Group), false);
-        assert_eq!(file.can_write(PosixPexQuery::Group), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), false);
-        assert_eq!(file.can_read(PosixPexQuery::Others), false);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), false);
+        let file: File = ListParser::parse_posix(
+            "drws------    2 u-redacted g-redacted      3864 Feb 17  2023 sas",
+        )
+        .unwrap();
+        pretty_assertions::assert_eq!(file.is_directory(), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), false);
+        let file: File = ListParser::parse_posix(
+            "drwS------    2 u-redacted g-redacted      3864 Feb 17  2023 sas",
+        )
+        .unwrap();
+        pretty_assertions::assert_eq!(file.is_directory(), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), false);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), false);
         // Setgid bit
-        let file: File =
-            File::from_str("drwx--s---    2 u-redacted g-redacted      3864 Feb 17  2023 sas")
-                .ok()
-                .unwrap();
-        assert_eq!(file.is_directory(), true);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), false);
-        assert_eq!(file.can_write(PosixPexQuery::Group), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), true);
-        assert_eq!(file.can_read(PosixPexQuery::Others), false);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), false);
-        let file: File =
-            File::from_str("drwx--S---    2 u-redacted g-redacted      3864 Feb 17  2023 sas")
-                .ok()
-                .unwrap();
-        assert_eq!(file.is_directory(), true);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), false);
-        assert_eq!(file.can_write(PosixPexQuery::Group), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), false);
-        assert_eq!(file.can_read(PosixPexQuery::Others), false);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), false);
+        let file: File = ListParser::parse_posix(
+            "drwx--s---    2 u-redacted g-redacted      3864 Feb 17  2023 sas",
+        )
+        .unwrap();
+        pretty_assertions::assert_eq!(file.is_directory(), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), false);
+        let file: File = ListParser::parse_posix(
+            "drwx--S---    2 u-redacted g-redacted      3864 Feb 17  2023 sas",
+        )
+        .unwrap();
+        pretty_assertions::assert_eq!(file.is_directory(), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), false);
         // Sticky bit
-        let file: File =
-            File::from_str("drwx-----t    2 u-redacted g-redacted      3864 Feb 17  2023 sas")
-                .ok()
-                .unwrap();
-        assert_eq!(file.is_directory(), true);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), false);
-        assert_eq!(file.can_write(PosixPexQuery::Group), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), false);
-        assert_eq!(file.can_read(PosixPexQuery::Others), false);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), true);
-        let file: File =
-            File::from_str("drwx--S--T    2 u-redacted g-redacted      3864 Feb 17  2023 sas")
-                .ok()
-                .unwrap();
-        assert_eq!(file.is_directory(), true);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), false);
-        assert_eq!(file.can_write(PosixPexQuery::Group), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), false);
-        assert_eq!(file.can_read(PosixPexQuery::Others), false);
-        assert_eq!(file.can_write(PosixPexQuery::Others), false);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), false);
+        let file: File = ListParser::parse_posix(
+            "drwx-----t    2 u-redacted g-redacted      3864 Feb 17  2023 sas",
+        )
+        .unwrap();
+        pretty_assertions::assert_eq!(file.is_directory(), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), true);
+        let file: File = ListParser::parse_posix(
+            "drwx--S--T    2 u-redacted g-redacted      3864 Feb 17  2023 sas",
+        )
+        .unwrap();
+        pretty_assertions::assert_eq!(file.is_directory(), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), false);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), false);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), false);
 
         // Error
-        assert_eq!(
-            File::from_posix_line("drwxrwxr-x 1 0  9  Nov 5 2018 docs")
-                .err()
-                .unwrap(),
+        pretty_assertions::assert_eq!(
+            ListParser::parse_posix("drwxrwxr-x 1 0  9  Nov 5 2018 docs").unwrap_err(),
             ParseError::SyntaxError
         );
-        assert_eq!(
-            File::from_posix_line("drwxrwxr-x 1 root  dialout  4096 Nov 31 2018 provola")
-                .err()
-                .unwrap(),
+        pretty_assertions::assert_eq!(
+            ListParser::parse_posix("drwxrwxr-x 1 root  dialout  4096 Nov 31 2018 provola")
+                .unwrap_err(),
             ParseError::InvalidDate
         );
     }
@@ -782,31 +569,31 @@ mod test {
     #[test]
     fn should_parse_utf8_names_in_ls_output() {
         assert!(
-            File::try_from("-rw-rw-r-- 1 омар  www-data  8192 Nov 5 2018 фообар.txt".to_string())
+            ListParser::parse_posix("-rw-rw-r-- 1 омар  www-data  8192 Nov 5 2018 фообар.txt")
                 .is_ok()
         );
     }
 
     #[test]
     fn parse_dos_line() {
-        let file: File = File::try_from("04-08-14  03:09PM  8192 omar.txt".to_string())
+        let file: File = ListParser::parse_dos("04-08-14  03:09PM  8192 omar.txt")
             .ok()
             .unwrap();
-        assert_eq!(file.name(), "omar.txt");
-        assert_eq!(file.size, 8192);
+        pretty_assertions::assert_eq!(file.name(), "omar.txt");
+        pretty_assertions::assert_eq!(file.size, 8192);
         assert!(file.is_file());
-        assert_eq!(file.gid, None);
-        assert_eq!(file.uid, None);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), true);
-        assert_eq!(file.can_write(PosixPexQuery::Group), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), true);
-        assert_eq!(file.can_read(PosixPexQuery::Others), true);
-        assert_eq!(file.can_write(PosixPexQuery::Others), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), true);
-        assert_eq!(
+        pretty_assertions::assert_eq!(file.gid, None);
+        pretty_assertions::assert_eq!(file.uid, None);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(
             file.modified
                 .duration_since(SystemTime::UNIX_EPOCH)
                 .ok()
@@ -814,23 +601,21 @@ mod test {
             Duration::from_secs(1396969740)
         );
         // Parse directory
-        let dir: File = File::try_from("04-08-14  03:09PM  <DIR> docs")
-            .ok()
-            .unwrap();
-        assert_eq!(dir.name(), "docs");
+        let dir: File = ListParser::parse_dos("04-08-14  03:09PM  <DIR> docs").unwrap();
+        pretty_assertions::assert_eq!(dir.name(), "docs");
         assert!(dir.is_directory());
-        assert_eq!(dir.uid, None);
-        assert_eq!(dir.gid, None);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), true);
-        assert_eq!(file.can_write(PosixPexQuery::Group), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), true);
-        assert_eq!(file.can_read(PosixPexQuery::Others), true);
-        assert_eq!(file.can_write(PosixPexQuery::Others), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), true);
-        assert_eq!(
+        pretty_assertions::assert_eq!(dir.uid, None);
+        pretty_assertions::assert_eq!(dir.gid, None);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(
             dir.modified
                 .duration_since(SystemTime::UNIX_EPOCH)
                 .ok()
@@ -838,20 +623,20 @@ mod test {
             Duration::from_secs(1396969740)
         );
         // Error
-        assert_eq!(
-            File::from_dos_line("-08-14  03:09PM  <DIR> docs")
+        pretty_assertions::assert_eq!(
+            ListParser::parse_dos("-08-14  03:09PM  <DIR> docs")
                 .err()
                 .unwrap(),
             ParseError::SyntaxError
         );
-        assert_eq!(
-            File::from_dos_line("34-08-14  03:09PM  <DIR> docs")
+        pretty_assertions::assert_eq!(
+            ListParser::parse_dos("34-08-14  03:09PM  <DIR> docs")
                 .err()
                 .unwrap(),
             ParseError::InvalidDate
         );
-        assert_eq!(
-            File::from_dos_line("04-08-14  03:09PM  OMAR docs")
+        pretty_assertions::assert_eq!(
+            ListParser::parse_dos("04-08-14  03:09PM  OMAR docs")
                 .err()
                 .unwrap(),
             ParseError::BadSize
@@ -860,13 +645,13 @@ mod test {
 
     #[test]
     fn test_should_parse_name_starting_with_tricky_numbers() {
-        let file = File::from_posix_line(
+        let file = ListParser::parse_posix(
             "-r--r--r--    1 23        23         1234567 Jan 1  2000 01 1234 foo.mp3",
         )
         .unwrap();
-        assert_eq!(file.name(), "01 1234 foo.mp3");
-        assert_eq!(file.size, 1234567);
-        assert_eq!(
+        pretty_assertions::assert_eq!(file.name(), "01 1234 foo.mp3");
+        pretty_assertions::assert_eq!(file.size, 1234567);
+        pretty_assertions::assert_eq!(
             file.modified
                 .duration_since(SystemTime::UNIX_EPOCH)
                 .ok()
@@ -876,23 +661,11 @@ mod test {
     }
 
     #[test]
-    fn get_name_and_link() {
-        assert_eq!(
-            File::get_name_and_link("Cargo.toml"),
-            (String::from("Cargo.toml"), None)
-        );
-        assert_eq!(
-            File::get_name_and_link("Cargo -> Cargo.toml"),
-            (String::from("Cargo"), Some(PathBuf::from("Cargo.toml")))
-        );
-    }
-
-    #[test]
     fn parse_lstime() {
         // Good cases
-        assert_eq!(
+        pretty_assertions::assert_eq!(
             fmt_time(
-                File::parse_lstime("Nov 5 16:32", "%b %d %Y", "%b %d %H:%M")
+                ListParser::parse_lstime("Nov 5 16:32", "%b %d %Y", "%b %d %H:%M")
                     .ok()
                     .unwrap(),
                 "%m %d %M"
@@ -900,9 +673,9 @@ mod test {
             .as_str(),
             "11 05 32"
         );
-        assert_eq!(
+        pretty_assertions::assert_eq!(
             fmt_time(
-                File::parse_lstime("Dec 2 21:32", "%b %d %Y", "%b %d %H:%M")
+                ListParser::parse_lstime("Dec 2 21:32", "%b %d %Y", "%b %d %H:%M")
                     .ok()
                     .unwrap(),
                 "%m %d %M"
@@ -910,8 +683,8 @@ mod test {
             .as_str(),
             "12 02 32"
         );
-        assert_eq!(
-            File::parse_lstime("Nov 5 2018", "%b %d %Y", "%b %d %H:%M")
+        pretty_assertions::assert_eq!(
+            ListParser::parse_lstime("Nov 5 2018", "%b %d %Y", "%b %d %H:%M")
                 .ok()
                 .unwrap()
                 .duration_since(SystemTime::UNIX_EPOCH)
@@ -919,8 +692,8 @@ mod test {
                 .unwrap(),
             Duration::from_secs(1541376000)
         );
-        assert_eq!(
-            File::parse_lstime("Mar 18 2018", "%b %d %Y", "%b %d %H:%M")
+        pretty_assertions::assert_eq!(
+            ListParser::parse_lstime("Mar 18 2018", "%b %d %Y", "%b %d %H:%M")
                 .ok()
                 .unwrap()
                 .duration_since(SystemTime::UNIX_EPOCH)
@@ -929,15 +702,15 @@ mod test {
             Duration::from_secs(1521331200)
         );
         // bad cases
-        assert!(File::parse_lstime("Oma 31 2018", "%b %d %Y", "%b %d %H:%M").is_err());
-        assert!(File::parse_lstime("Feb 31 2018", "%b %d %Y", "%b %d %H:%M").is_err());
-        assert!(File::parse_lstime("Feb 15 25:32", "%b %d %Y", "%b %d %H:%M").is_err());
+        assert!(ListParser::parse_lstime("Oma 31 2018", "%b %d %Y", "%b %d %H:%M").is_err());
+        assert!(ListParser::parse_lstime("Feb 31 2018", "%b %d %Y", "%b %d %H:%M").is_err());
+        assert!(ListParser::parse_lstime("Feb 15 25:32", "%b %d %Y", "%b %d %H:%M").is_err());
     }
 
     #[test]
     fn parse_dostime() {
-        assert_eq!(
-            File::parse_dostime("04-08-14  03:09PM")
+        pretty_assertions::assert_eq!(
+            ListParser::parse_dostime("04-08-14  03:09PM")
                 .ok()
                 .unwrap()
                 .duration_since(SystemTime::UNIX_EPOCH)
@@ -946,88 +719,56 @@ mod test {
             Duration::from_secs(1396969740)
         );
         // Not enough argument for datetime
-        assert!(File::parse_dostime("04-08-14").is_err());
+        assert!(ListParser::parse_dostime("04-08-14").is_err());
     }
 
     #[test]
     fn test_parse_mlsx_line() {
-        let file = File::from_mlsx_line("type=file;size=8192;modify=20181105163248; omar.txt")
-            .ok()
-            .unwrap();
+        let file =
+            ListParser::parse_mlsd("type=file;size=8192;modify=20181105163248; omar.txt").unwrap();
 
-        assert_eq!(file.name(), "omar.txt");
-        assert_eq!(file.size, 8192);
+        pretty_assertions::assert_eq!(file.name(), "omar.txt");
+        pretty_assertions::assert_eq!(file.size, 8192);
         assert!(file.is_file());
-        assert_eq!(file.gid, None);
-        assert_eq!(file.uid, None);
-        assert_eq!(file.can_read(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_write(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
-        assert_eq!(file.can_read(PosixPexQuery::Group), true);
-        assert_eq!(file.can_write(PosixPexQuery::Group), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Group), true);
-        assert_eq!(file.can_read(PosixPexQuery::Others), true);
-        assert_eq!(file.can_write(PosixPexQuery::Others), true);
-        assert_eq!(file.can_execute(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.gid, None);
+        pretty_assertions::assert_eq!(file.uid, None);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Owner), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Group), true);
+        pretty_assertions::assert_eq!(file.can_read(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.can_write(PosixPexQuery::Others), true);
+        pretty_assertions::assert_eq!(file.can_execute(PosixPexQuery::Others), true);
 
-        let file = File::from_mlsx_line("type=dir;size=4096;modify=20181105163248; docs")
-            .ok()
-            .unwrap();
+        let file =
+            ListParser::parse_mlsd("type=dir;size=4096;modify=20181105163248; docs").unwrap();
 
-        assert_eq!(file.name(), "docs");
+        pretty_assertions::assert_eq!(file.name(), "docs");
         assert!(file.is_directory());
 
-        let file = File::from_mlsx_line(
+        let file = ListParser::parse_mlsd(
             "type=file;size=4096;modify=20181105163248;unix.mode=644; omar.txt",
         )
-        .ok()
         .unwrap();
-        assert_eq!(
+        pretty_assertions::assert_eq!(
             file.posix_pex,
             (PosixPex::from(6), PosixPex::from(4), PosixPex::from(4))
         );
     }
 
     #[test]
-    fn file_type() {
-        assert_eq!(FileType::Directory.is_directory(), true);
-        assert_eq!(FileType::Directory.is_file(), false);
-        assert_eq!(FileType::Directory.is_symlink(), false);
-        assert_eq!(FileType::Directory.symlink(), None);
-        assert_eq!(FileType::File.is_directory(), false);
-        assert_eq!(FileType::File.is_file(), true);
-        assert_eq!(FileType::File.is_symlink(), false);
-        assert_eq!(FileType::File.symlink(), None);
-        assert_eq!(FileType::Symlink(PathBuf::default()).is_directory(), false);
-        assert_eq!(FileType::Symlink(PathBuf::default()).is_file(), false);
-        assert_eq!(FileType::Symlink(PathBuf::default()).is_symlink(), true);
-        assert_eq!(
-            FileType::Symlink(PathBuf::default()).symlink(),
-            Some(PathBuf::default().as_path())
+    fn get_name_and_link() {
+        pretty_assertions::assert_eq!(
+            ListParser::get_name_and_link("Cargo.toml"),
+            (String::from("Cargo.toml"), None)
+        );
+        pretty_assertions::assert_eq!(
+            ListParser::get_name_and_link("Cargo -> Cargo.toml"),
+            (String::from("Cargo"), Some(PathBuf::from("Cargo.toml")))
         );
     }
-
-    #[test]
-    fn posix_pex_from_bits() {
-        let pex: PosixPex = PosixPex::from(4);
-        assert_eq!(pex.can_read(), true);
-        assert_eq!(pex.can_write(), false);
-        assert_eq!(pex.can_execute(), false);
-        let pex: PosixPex = PosixPex::from(0);
-        assert_eq!(pex.can_read(), false);
-        assert_eq!(pex.can_write(), false);
-        assert_eq!(pex.can_execute(), false);
-        let pex: PosixPex = PosixPex::from(3);
-        assert_eq!(pex.can_read(), false);
-        assert_eq!(pex.can_write(), true);
-        assert_eq!(pex.can_execute(), true);
-        let pex: PosixPex = PosixPex::from(7);
-        assert_eq!(pex.can_read(), true);
-        assert_eq!(pex.can_write(), true);
-        assert_eq!(pex.can_execute(), true);
-    }
-
-    // -- utils
 
     fn fmt_time(time: SystemTime, fmt: &str) -> String {
         let datetime: DateTime<Utc> = time.into();
